refactor: reviews changes
